backend/requirements.txt
fastapi
uvicorn
openai-whisper
transformers
spacy
python-multipart
pydantic
